:zap: remove int64

+++ deepchem/feat/tests/test_graph_data.py
@@ -11,11 +11,10 @@ class TestGraph(unittest.TestCase):
     num_edges, num_edge_features = 6, 32
     node_features = np.random.random_sample((num_nodes, num_node_features))
     edge_features = np.random.random_sample((num_edges, num_edge_features))
-    edge_index = np.array(
-        [
-            [0, 1, 2, 2, 3, 4],
-            [1, 2, 0, 3, 4, 0],
-        ], dtype=np.int64)
+    edge_index = np.array([
+        [0, 1, 2, 2, 3, 4],
+        [1, 2, 0, 3, 4, 0],
+    ])
     graph_features = None
 
     graph = GraphData(
@@ -41,11 +40,10 @@ class TestGraph(unittest.TestCase):
   def test_invalid_graph_data(self):
     with pytest.raises(ValueError):
       invalid_node_features_type = list(np.random.random_sample((5, 32)))
-      edge_index = np.array(
-          [
-              [0, 1, 2, 2, 3, 4],
-              [1, 2, 0, 3, 4, 0],
-          ], dtype=np.int64)
+      edge_index = np.array([
+          [0, 1, 2, 2, 3, 4],
+          [1, 2, 0, 3, 4, 0],
+      ])
       _ = GraphData(
           node_features=invalid_node_features_type,
           edge_index=edge_index,
@@ -53,11 +51,10 @@ class TestGraph(unittest.TestCase):
 
     with pytest.raises(ValueError):
       node_features = np.random.random_sample((5, 32))
-      invalid_edge_index_shape = np.array(
-          [
-              [0, 1, 2, 2, 3, 4],
-              [1, 2, 0, 3, 4, 5],
-          ], dtype=np.int64)
+      invalid_edge_index_shape = np.array([
+          [0, 1, 2, 2, 3, 4],
+          [1, 2, 0, 3, 4, 5],
+      ])
       _ = GraphData(
           node_features=node_features,
           edge_index=invalid_edge_index_shape,
@@ -65,13 +62,11 @@ class TestGraph(unittest.TestCase):
 
     with pytest.raises(ValueError):
       node_features = np.random.random_sample((5, 5))
-      invalid_edge_index_shape = np.array(
-          [
-              [0, 1, 2, 2, 3, 4],
-              [1, 2, 0, 3, 4, 0],
-              [2, 2, 1, 4, 0, 3],
-          ],
-          dtype=np.int64)
+      invalid_edge_index_shape = np.array([
+          [0, 1, 2, 2, 3, 4],
+          [1, 2, 0, 3, 4, 0],
+          [2, 2, 1, 4, 0, 3],
+      ],)
       _ = GraphData(
           node_features=node_features,
           edge_index=invalid_edge_index_shape,
@@ -85,9 +80,9 @@ class TestGraph(unittest.TestCase):
     num_nodes_list, num_edge_list = [3, 4, 5], [2, 4, 5]
     num_node_features, num_edge_features = 32, 32
     edge_index_list = [
-        np.array([[0, 1], [1, 2]], dtype=np.int64),
-        np.array([[0, 1, 2, 3], [1, 2, 0, 2]], dtype=np.int64),
-        np.array([[0, 1, 2, 3, 4], [1, 2, 3, 4, 0]], dtype=np.int64),
+        np.array([[0, 1], [1, 2]]),
+        np.array([[0, 1, 2, 3], [1, 2, 0, 2]]),
+        np.array([[0, 1, 2, 3, 4], [1, 2, 3, 4, 0]]),
     ]
 
     graph_list = [
